Refactored code to put non-route functions to separate file
